chore: factor methods into lib/
