+++ pages/api/nftInfo/[uri].ts
@@ -1,11 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import IPFSGatewayTools from '@pinata/ipfs-gateway-tools/dist/node';
 import { captureException, withSentry } from '@sentry/nextjs';
-import { getMimeType } from 'lib/getNFTInfo';
-
-const ipfsGatewayTools = new IPFSGatewayTools();
-
-type Media = { mediaUrl: string; mediaMimeType: string } | null;
+import { convertIPFS, getMedia, Media } from 'lib/getNFTInfo';
 
 export type NFTResponseData = {
   name: string;
@@ -40,24 +35,13 @@ async function handler(
       external_url,
     } = await tokenURIRes.json();
 
-    const finalImage = convertIPFS(image || image_url);
-    const finalAnimation = convertIPFS(animation_url);
-
-    const [animationMimeType, imageMimeType] = await Promise.all([
-      getMimeType(animation_url || ''),
-      getMimeType(image || ''),
-    ]);
+    const media = await getMedia({ animation_url, image, image_url });
 
     res.status(200).json({
       name,
       description,
       tokenId,
-      image: finalImage
-        ? { mediaUrl: finalImage, mediaMimeType: imageMimeType }
-        : null,
-      animation: finalAnimation
-        ? { mediaUrl: finalAnimation, mediaMimeType: animationMimeType }
-        : null,
+      ...media,
       external_url,
     });
   } catch (e) {
@@ -74,31 +58,6 @@ async function handler(
     // TODO: we could respond with a failure reason and surface that in the UI.
     return res.status(404).json(null);
   }
-}
-
-/**
- * If `uri` is an IPFS uri, convert to use our gateway. Otherwise return it untouched.
- */
-function convertIPFS(uri?: string): string | undefined {
-  if (!uri) {
-    return uri;
-  }
-
-  try {
-    if (ipfsGatewayTools.containsCID(uri).containsCid) {
-      return ipfsGatewayTools.convertToDesiredGateway(
-        uri,
-        'https://nftpawnshop.mypinata.cloud',
-      );
-    }
-  } catch (e) {
-    if ((e as any).message !== 'url is not string') {
-      // got an annoying false positive message here; uri would definitely be
-      // type string but the error would be raised anyway. Skipping those.
-      captureException(e);
-    }
-  }
-  return uri;
 }
 
 export default withSentry(handler);
